Created SimulationInformation output object for more control (refs #4454)
