+++ framework/src/base/CoupledProblem.C
@@ -35,8 +35,3 @@ void
 CoupledProblem::init()
 {
 }
-
-void
-CoupledProblem::output(bool /*force*/)
-{
-}
