+++ framework/src/outputs/Console.C
@@ -30,7 +30,7 @@ InputParameters validParams<Console>()
 
   // Get the parameters from the base class
   InputParameters params = validParams<TableOutput>();
-  params += TableOutput::enableOutputTypes("system_information scalar postprocessor input");
+  params += TableOutput::enableOutputTypes("scalar postprocessor input");
 
   // Screen and file output toggles
   params.addParam<bool>("output_screen", true, "Output to the screen");
@@ -43,7 +43,6 @@ InputParameters validParams<Console>()
 
   // Verbosity
   params.addParam<bool>("verbose", false, "Print detailed diagnostics on timestep calculation");
-  params.addParam<bool>("show_output_on", false, "Print the output execution with the system information");
 
   // Basic table output controls
   params.addParam<bool>("scientific_time", false, "Control the printing of time and dt in scientific notation");
@@ -88,9 +87,6 @@ InputParameters validParams<Console>()
    * the set method. This enables "quiet mode". This is done to allow for the proper detection
    * of user-modified parameters
    */
-  // By default set System Information to output on initial
-  params.set<MultiMooseEnum>("output_system_information_on", /*quiet_mode=*/true) = "initial";
-
   // Change the default behavior of 'output_on' to included nonlinear iterations and failed timesteps
   params.set<MultiMooseEnum>("output_on", /*quiet_mode=*/true).push_back("nonlinear failed");
 
@@ -122,7 +118,6 @@ Console::Console(const std::string & name, InputParameters parameters) :
     _outlier_variable_norms(getParam<bool>("outlier_variable_norms")),
     _outlier_multiplier(getParam<std::vector<Real> >("outlier_multiplier")),
     _precision(isParamValid("time_precision") ? getParam<unsigned int>("time_precision") : 0),
-    _show_output_on_info(getParam<bool>("show_output_on")),
     _timing(_app.getParam<bool>("timing")),
     _console_buffer(_app.getOutputWarehouse().consoleBuffer()),
     _old_linear_norm(std::numeric_limits<Real>::max()),
@@ -168,10 +163,6 @@ Console::Console(const std::string & name, InputParameters parameters) :
         Moose::_color_console = false;
     }
   }
-
-  // If --show-outputs is used, enable it
-  if (_app.getParam<bool>("show_outputs"))
-    _show_output_on_info = true;
 }
 
 Console::~Console()
@@ -255,11 +246,6 @@ Console::output(const ExecFlagType & type)
   // Flush the Console buffer, if we don't do this here then the linear/nonlinear residual output
   // may write to the screen prior to buffered text
   _app.getOutputWarehouse().flushConsoleBuffer();
-
-  // Output the system information first; this forces this to be the first item to write by default
-  // However, 'output_system_information_on' still operates correctly, so it may be changed by the user
-  if (shouldOutput("system_information", type))
-    outputSystemInformation();
 
   // Write the input
   if (shouldOutput("input", type))
@@ -471,19 +457,6 @@ Console::outputNorm(const Real & old_norm, const Real & norm)
   return oss.str();
 }
 
-
-// Method for stringstream formatting
-void
-Console::insertNewline(std::stringstream &oss, std::streampos &begin, std::streampos &curr)
-{
-   if (curr - begin > _line_length)
-   {
-     oss << "\n";
-     begin = oss.tellp();
-     oss << std::setw(_field_width + 2) << "";  // "{ "
-   }
-}
-
 void
 Console::outputInput()
 {
@@ -573,198 +546,6 @@ Console::mooseConsole(const std::string & message)
 
   // Flush the stream to the screen
   Moose::out << std::flush;
-}
-
-void
-Console::outputSystemInformation()
-{
-  // Don't build this information if nothing is to be written
-  if (!_write_screen && !_write_file)
-    return;
-
-  std::stringstream oss;
-
-  // Framework information
-  if (_app.getSystemInfo() != NULL)
-    oss << _app.getSystemInfo()->getInfo();
-
-  if (_problem_ptr->legacyUoAuxComputation() || _problem_ptr->legacyUoInitialization())
-  {
-    oss << "LEGACY MODES ENABLED:\n";
-    if (_problem_ptr->legacyUoAuxComputation())
-      oss << "  Computing EXEC_LINEAR AuxKernel types when any UserObject type is executed.\n";
-    if (_problem_ptr->legacyUoInitialization())
-      oss << "  Computing all UserObjects during initial setup.\n";
-  }
-
-  oss << std::left << '\n'
-      << "Parallelism:\n"
-      << std::setw(_field_width) << "  Num Processors: " << static_cast<std::size_t>(n_processors()) << '\n'
-      << std::setw(_field_width) << "  Num Threads: " << static_cast<std::size_t>(n_threads()) << '\n'
-      << '\n';
-
-  MooseMesh & moose_mesh = _problem_ptr->mesh();
-  MeshBase & mesh = moose_mesh.getMesh();
-  oss << "Mesh: " << '\n'
-      << std::setw(_field_width) << "  Distribution: " << (moose_mesh.isParallelMesh() ? "parallel" : "serial")
-      << (moose_mesh.isDistributionForced() ? " (forced) " : "") << '\n'
-      << std::setw(_field_width) << "  Mesh Dimension: " << mesh.mesh_dimension() << '\n'
-      << std::setw(_field_width) << "  Spatial Dimension: " << mesh.spatial_dimension() << '\n'
-      << std::setw(_field_width) << "  Nodes:" << '\n'
-      << std::setw(_field_width) << "    Total:" << mesh.n_nodes() << '\n'
-      << std::setw(_field_width) << "    Local:" << mesh.n_local_nodes() << '\n'
-      << std::setw(_field_width) << "  Elems:" << '\n'
-      << std::setw(_field_width) << "    Total:" << mesh.n_elem() << '\n'
-      << std::setw(_field_width) << "    Local:" << mesh.n_local_elem() << '\n'
-      << std::setw(_field_width) << "  Num Subdomains: "       << static_cast<std::size_t>(mesh.n_subdomains()) << '\n'
-      << std::setw(_field_width) << "  Num Partitions: "       << static_cast<std::size_t>(mesh.n_partitions()) << '\n';
-  if (n_processors() > 1 && moose_mesh.partitionerName() != "")
-    oss << std::setw(_field_width) << "  Partitioner: "       << moose_mesh.partitionerName()
-        << (moose_mesh.isPartitionerForced() ? " (forced) " : "")
-        << '\n';
-  oss << '\n';
-
-  EquationSystems & eq = _problem_ptr->es();
-  unsigned int num_systems = eq.n_systems();
-  for (unsigned int i=0; i<num_systems; ++i)
-  {
-    const System & system = eq.get_system(i);
-    if (system.system_type() == "TransientNonlinearImplicit")
-      oss << "Nonlinear System:" << '\n';
-    else if (system.system_type() == "TransientExplicit")
-      oss << "Auxiliary System:" << '\n';
-    else
-      oss << std::setw(_field_width) << system.system_type() << '\n';
-
-    if (system.n_dofs())
-    {
-      oss << std::setw(_field_width) << "  Num DOFs: " << system.n_dofs() << '\n'
-          << std::setw(_field_width) << "  Num Local DOFs: " << system.n_local_dofs() << '\n';
-
-      std::streampos begin_string_pos = oss.tellp();
-      std::streampos curr_string_pos = begin_string_pos;
-      oss << std::setw(_field_width) << "  Variables: ";
-      for (unsigned int vg=0; vg<system.n_variable_groups(); vg++)
-      {
-        const VariableGroup &vg_description (system.variable_group(vg));
-
-        if (vg_description.n_variables() > 1) oss << "{ ";
-        for (unsigned int vn=0; vn<vg_description.n_variables(); vn++)
-        {
-          oss << "\"" << vg_description.name(vn) << "\" ";
-          curr_string_pos = oss.tellp();
-          insertNewline(oss, begin_string_pos, curr_string_pos);
-        }
-
-        if (vg_description.n_variables() > 1) oss << "} ";
-      }
-      oss << '\n';
-
-      begin_string_pos = oss.tellp();
-      curr_string_pos = begin_string_pos;
-      oss << std::setw(_field_width) << "  Finite Element Types: ";
-#ifndef LIBMESH_ENABLE_INFINITE_ELEMENTS
-      for (unsigned int vg=0; vg<system.n_variable_groups(); vg++)
-      {
-        oss << "\""
-            << libMesh::Utility::enum_to_string<FEFamily>(system.get_dof_map().variable_group(vg).type().family)
-            << "\" ";
-        curr_string_pos = oss.tellp();
-        insertNewline(oss, begin_string_pos, curr_string_pos);
-      }
-      oss << '\n';
-#else
-      for (unsigned int vg=0; vg<system.n_variable_groups(); vg++)
-      {
-        oss << "\""
-            << libMesh::Utility::enum_to_string<FEFamily>(system.get_dof_map().variable_group(vg).type().family)
-            << "\", \""
-            << libMesh::Utility::enum_to_string<FEFamily>(system.get_dof_map().variable_group(vg).type().radial_family)
-            << "\" ";
-        curr_string_pos = oss.tellp();
-        insertNewline(oss, begin_string_pos, curr_string_pos);
-      }
-      oss << '\n';
-
-      begin_string_pos = oss.tellp();
-      curr_string_pos = begin_string_pos;
-      oss << std::setw(_field_width) << "  Infinite Element Mapping: ";
-      for (unsigned int vg=0; vg<system.n_variable_groups(); vg++)
-      {
-        oss << "\""
-            << libMesh::Utility::enum_to_string<InfMapType>(system.get_dof_map().variable_group(vg).type().inf_map)
-            << "\" ";
-        curr_string_pos = oss.tellp();
-        insertNewline(oss, begin_string_pos, curr_string_pos);
-      }
-      oss << '\n';
-#endif
-
-      begin_string_pos = oss.tellp();
-      curr_string_pos = begin_string_pos;
-      oss << std::setw(_field_width) << "  Approximation Orders: ";
-      for (unsigned int vg=0; vg<system.n_variable_groups(); vg++)
-      {
-#ifndef LIBMESH_ENABLE_INFINITE_ELEMENTS
-        oss << "\""
-            << Utility::enum_to_string<Order>(system.get_dof_map().variable_group(vg).type().order)
-            << "\" ";
-#else
-        oss << "\""
-            << Utility::enum_to_string<Order>(system.get_dof_map().variable_group(vg).type().order)
-            << "\", \""
-            << Utility::enum_to_string<Order>(system.get_dof_map().variable_group(vg).type().radial_order)
-            << "\" ";
-#endif
-        curr_string_pos = oss.tellp();
-        insertNewline(oss, begin_string_pos, curr_string_pos);
-      }
-      oss << "\n\n";
-    }
-    else
-      oss << "   *** EMPTY ***\n\n";
-  }
-
-  oss << "Execution Information:\n"
-      << std::setw(_field_width) << "  Executioner: " << demangle(typeid(*_app.getExecutioner()).name()) << '\n';
-
-  std::string time_stepper = _app.getExecutioner()->getTimeStepperName();
-  if (time_stepper != "")
-    oss << std::setw(_field_width) << "  TimeStepper: " << time_stepper << '\n';
-
-  oss << std::setw(_field_width) << "  Solver Mode: " << Moose::stringify<Moose::SolveType>(_problem_ptr->solverParams()._type) << '\n';
-
-  const std::string & pc_desc = _problem_ptr->getPreconditionerDescription();
-  if (!pc_desc.empty())
-    oss << std::setw(_field_width) << "  Preconditioner: " << pc_desc << '\n';
-  oss << '\n';
-
-  // Output information
-  if (_show_output_on_info)
-  {
-    const std::vector<Output *> & outputs = _app.getOutputWarehouse().all();
-    oss << "Outputs:\n";
-    for (std::vector<Output *>::const_iterator it = outputs.begin(); it != outputs.end(); ++it)
-    {
-      // Display the "output_on" settings
-      const MultiMooseEnum & output_on = (*it)->outputOn();
-      oss << "  " << std::setw(_field_width-2) << (*it)->name() <<  "\"" << output_on << "\"\n";
-
-      // Display the advanced "output_on" settings, only if they are different from "output_on"
-      if ((*it)->isAdvanced())
-      {
-        const OutputOnWarehouse & adv_on = (*it)->advancedOutputOn();
-        for (std::map<std::string, MultiMooseEnum>::const_iterator adv_it = adv_on.begin(); adv_it != adv_on.end(); ++adv_it)
-          if (output_on != adv_it->second)
-            oss << "    " << std::setw(_field_width-4) << adv_it->first + ":" <<  "\"" << adv_it->second << "\"\n";
-      }
-    }
-  }
-
-  oss << "\n\n";
-
-  // Output the information
-  write(oss.str());
 }
 
 void
